requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
pymongo==4.6.1
requests==2.31.0
dnspython==2.4.2
python-dateutil==2.8.2
geoip2==4.8.0
maxminddb-geolite2==2018.703
flask==3.0.0
psutil==5.9.8
gunicorn==21.2.0
urllib3==2.1.0
aiohttp==3.9.1
colorama==0.4.6
python-socketio==5.10.0
websockets==12.0 